# pkg/api/rest/docs/swagger.yaml
basePath: /cicada
definitions:
  common.ErrorResponse:
    properties:
      error:
        type: string
    type: object
...
  model.CreateDataReq:
    properties:
      description:
        type: string
      task_groups:
        items:
          $ref: '#/definitions/model.CreateTaskGroupReq'
        type: array
    required:
    - task_groups
    type: object
...
  model.CreateTaskGroupReq:
    properties:
      description:
        type: string
      name:
        type: string
      tasks:
        items:
          $ref: '#/definitions/model.CreateTaskReq'
        type: array
    required:
    - name
    - tasks
    type: object
  model.CreateTaskReq:
    properties:
      dependencies:
        items:
          type: string
        type: array
      extra:
        additionalProperties: true
        type: object
      name:
        type: string
      path_params:
        additionalProperties:
          type: string
        type: object
      query_params:
        additionalProperties:
          type: string
        type: object
      request_body:
        type: string
      task_component:
        type: string
    required:
    - name
    - request_body
    - task_component
    type: object
  model.CreateWorkflowReq:
    properties:
      data:
        $ref: '#/definitions/model.CreateDataReq'
      name:
        type: string
      spec_version:
        type: string
    required:
    - data
    - name
    type: object
...
  model.Data:
    properties:
      description:
        type: string
      task_groups:
        items:
          $ref: '#/definitions/model.TaskGroup'
        type: array
    required:
    - task_groups
    type: object
...
  model.SimpleMsg:
    properties:
      message:
        type: string
    type: object
...
  model.Task:
    properties:
      dependencies:
        items:
          type: string
        type: array
      extra:
        additionalProperties: true
        type: object
      id:
        type: string
      name:
        type: string
      path_params:
        additionalProperties:
          type: string
        type: object
      query_params:
        additionalProperties:
          type: string
        type: object
      request_body:
        type: string
      task_component:
        type: string
      task_group_id:
        type: string
      workflow_id:
        type: string
    required:
    - id
    - name
    - request_body
    - task_component
    - task_group_id
    - workflow_id
    type: object
...
  model.TaskGroup:
    properties:
      description:
        type: string
      id:
        type: string
      name:
        type: string
      tasks:
        items:
          $ref: '#/definitions/model.Task'
        type: array
      workflow_id:
        type: string
    required:
    - id
    - name
    - tasks
    - workflow_id
    type: object
...
  model.Workflow:
    properties:
      created_at:
        type: string
      data:
        $ref: '#/definitions/model.Data'
      id:
        type: string
      name:
        type: string
      spec_version:
        type: string
      updated_at:
        type: string
    required:
    - data
    - id
    - name
    - spec_version
    type: object
...
  model.WorkflowTemplate:
    properties:
      data:
        $ref: '#/definitions/model.Data'
      name:
        type: string
      spec_version:
        type: string
    required:
    - data
    - name
    - spec_version
    type: object
...
info:
  contact: {}
  description: Workflow management module
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: CM-Cicada REST API
  version: latest
paths:
...
  /workflow:
    get:
      consumes:
      - application/json
      description: Get a workflow list.
      operationId: list-workflow
      parameters:
      - description: Name of the workflow
        in: query
        name: name
        type: string
      - description: Page of the workflow list.
        in: query
        name: page
        type: string
      - description: Row of the workflow list.
        in: query
        name: row
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Successfully get a workflow list.
          schema:
            items:
              $ref: '#/definitions/model.Workflow'
            type: array
        "400":
          description: Sent bad request.
          schema:
            $ref: '#/definitions/common.ErrorResponse'
        "500":
          description: Failed to get a workflow list.
          schema:
            $ref: '#/definitions/common.ErrorResponse'
      summary: List Workflow
      tags:
      - '[Workflow]'
    post:
      consumes:
      - application/json
      description: Create a workflow.
      operationId: create-workflow
      parameters:
      - description: Workflow content
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.CreateWorkflowReq'
      produces:
      - application/json
      responses:
        "200":
          description: Successfully create the workflow.
          schema:
            $ref: '#/definitions/model.WorkflowTemplate'
        "400":
          description: Sent bad request.
          schema:
            $ref: '#/definitions/common.ErrorResponse'
        "500":
          description: Failed to create workflow.
          schema:
            $ref: '#/definitions/common.ErrorResponse'
      summary: Create Workflow
      tags:
      - '[Workflow]'
  /workflow/{wfId}:
    delete:
      consumes:
      - application/json
      description: Delete the workflow.
      operationId: delete-workflow
      parameters:
      - description: ID of the workflow.
        in: path
        name: wfId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Successfully delete the workflow
          schema:
            $ref: '#/definitions/model.SimpleMsg'
        "400":
          description: Sent bad request.
          schema:
            $ref: '#/definitions/common.ErrorResponse'
        "500":
          description: Failed to delete the workflow
          schema:
            $ref: '#/definitions/common.ErrorResponse'
      summary: Delete Workflow
      tags:
      - '[Workflow]'
    get:
      consumes:
      - application/json
      description: Get the workflow.
      operationId: get-workflow
      parameters:
      - description: ID of the workflow.
        in: path
        name: wfId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Successfully get the workflow.
          schema:
            $ref: '#/definitions/model.Workflow'
        "400":
          description: Sent bad request.
          schema:
            $ref: '#/definitions/common.ErrorResponse'
        "500":
          description: Failed to get the workflow.
          schema:
            $ref: '#/definitions/common.ErrorResponse'
      summary: Get Workflow
      tags:
      - '[Workflow]'
    put:
      consumes:
      - application/json
      description: Update the workflow content.
      operationId: update-workflow
      parameters:
      - description: ID of the workflow.
        in: path
        name: wfId
        required: true
        type: string
      - description: Workflow to modify.
        in: body
        name: Workflow
        required: true
        schema:
          $ref: '#/definitions/model.CreateWorkflowReq'
      produces:
      - application/json
      responses:
        "200":
          description: Successfully update the workflow
          schema:
            $ref: '#/definitions/model.Workflow'
        "400":
          description: Sent bad request.
          schema:
            $ref: '#/definitions/common.ErrorResponse'
        "500":
          description: Failed to update the workflow
          schema:
            $ref: '#/definitions/common.ErrorResponse'
      summary: Update Workflow
      tags:
      - '[Workflow]'
...
  /workflow/{wfId}/run:
    post:
      consumes:
      - application/json
      description: Run the workflow.
      operationId: run-workflow
      parameters:
      - description: ID of the workflow.
        in: path
        name: wfId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Successfully run the workflow.
          schema:
            $ref: '#/definitions/model.SimpleMsg'
        "400":
          description: Sent bad request.
          schema:
            $ref: '#/definitions/common.ErrorResponse'
        "500":
          description: Failed to run the Workflow
          schema:
            $ref: '#/definitions/common.ErrorResponse'
      summary: Run Workflow
      tags:
      - '[Workflow]'
...
  /workflow/name/{wfName}:
    get:
      consumes:
      - application/json
      description: Get the workflow by name.
      operationId: get-workflow-by-name
      parameters:
      - description: Name of the workflow.
        in: path
        name: wfName
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Successfully get the workflow.
          schema:
            $ref: '#/definitions/model.Workflow'
        "400":
          description: Sent bad request.
          schema:
            $ref: '#/definitions/common.ErrorResponse'
        "500":
          description: Failed to get the workflow.
          schema:
            $ref: '#/definitions/common.ErrorResponse'
      summary: Get Workflow by Name
      tags:
      - '[Workflow]'
  /workflow/tx:
    post:
      consumes:
      - application/json
      description: Create a workflow with TaskGroups/Tasks in a single DB transaction.
      operationId: create-workflow-tx
      parameters:
      - description: Workflow content
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.CreateWorkflowReq'
      produces:
      - application/json
      responses:
        "200":
          description: Successfully create the workflow.
          schema:
            $ref: '#/definitions/model.Workflow'
        "400":
          description: Sent bad request.
          schema:
            $ref: '#/definitions/common.ErrorResponse'
        "500":
          description: Failed to create workflow.
          schema:
            $ref: '#/definitions/common.ErrorResponse'
      summary: Create Workflow (DB Tx)
      tags:
      - '[Workflow]'
...
swagger: "2.0"